requirements.txt
requests
beautifulsoup4
spacy
flask
flask-cors
google-generativeai